[UPDATE] Socket connection in ChatActivity
