+++ app/src/main/java/com/example/project_flow_android/ui/chat/ChatActivity.kt
@@ -5,15 +5,20 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import com.example.project_flow_android.R
 import com.example.project_flow_android.ui.chat.fragment.ChatListFragment
+import com.example.project_flow_android.viewmodel.chat.ChatViewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class ChatActivity : AppCompatActivity() {
-
+    private val chatViewModel : ChatViewModel by viewModel()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_chat)
 
         val fragmentTransaction = supportFragmentManager.beginTransaction()
         fragmentTransaction.replace(R.id.chat_frame_container, ChatListFragment()).commit()
+
+        chatViewModel.connect()
+        chatViewModel.subscribe("/${chatViewModel.getProjectId()}/chatroom")
     }
 
     fun replace(fragment: Fragment) {
@@ -21,5 +26,10 @@ class ChatActivity : AppCompatActivity() {
         fragmentTransaction.setCustomAnimations(R.anim.from_bottom, R.anim.to_top, R.anim.from_top, R.anim.to_bottom)
             .addToBackStack(null)
         fragmentTransaction.replace(R.id.chat_frame_container, fragment).commit()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        chatViewModel.disconnect()
     }
 }
