+++ app/src/main/java/com/example/project_flow_android/ui/chat/fragment/ChatListFragment.kt
@@ -12,7 +12,6 @@ import kotlinx.android.synthetic.main.fragment_chat_list.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class ChatListFragment : Fragment() {
-    private val chatViewModel : ChatViewModel by viewModel()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -25,15 +24,8 @@ class ChatListFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        chatViewModel.connect()
-
         chat_list_add_iv.setOnClickListener{
             (activity as ChatActivity).replace(ChatCreateFragment())
         }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        chatViewModel.disconnect()
     }
 }
